make std::intrinsic functions actually be intrinsics
